fix fetch posts when needed while WS reconnects
